Completed merge of actor helper functions
